+++ .prettierrc.json
@@ -1,7 +1,7 @@
 {
   "endOfLine": "lf",
   "useTabs": false,
-  "printWidth": 120,
+  "printWidth": 160,
   "tabWidth": 2,
   "singleQuote": true,
   "bracketSpacing": true,
