+++ website/gatsby-node.js
@@ -17,6 +17,7 @@ exports.onCreateWebpackConfig = ({ stage, actions }) => {
 // You can delete this file if you're not using it
 const path = require('path');
 const docPageTemplate = path.resolve('src/templates/docPage.tsx');
+const frontPageTemplate = path.resolve('src/templates/frontpage.tsx');
 const sideBar = require('./config/sidebar.json');
 
 exports.createPages = async ({ graphql, actions }) => {
@@ -42,6 +43,9 @@ exports.createPages = async ({ graphql, actions }) => {
 
   const posts = result.data.allMarkdownRemark.edges;
   const idTitleMap = {};
+  const languages = [];
+
+  // create documentation pages
   posts.forEach(({ node }) => {
     console.log('-node.fileAbsolutePath-', node.frontmatter);
     const parsedPath = path.parse(node.fileAbsolutePath);
@@ -50,9 +54,10 @@ exports.createPages = async ({ graphql, actions }) => {
     const name = parsedPath.name;
     const title = node.frontmatter.title;
     const lng = parsedPath.dir.match('translated_docs') ? parsedPath.dir.split('/').pop() : 'en';
-    console.log('-lng', lng);
-    console.log('-markDownId', markDownId);
-    console.log('-title', title);
+    if (!languages.includes(lng)) {
+      languages.push(lng);
+    }
+
     if (!idTitleMap[lng]) {
       idTitleMap[lng] = {};
     }
@@ -63,6 +68,15 @@ exports.createPages = async ({ graphql, actions }) => {
       path: `docs/${lng}/${name}.html`,
       component: docPageTemplate,
       context: { id, lng, sideBar, title, idTitleMap, markDownId },
+    });
+  });
+
+  // create index pages
+  languages.map((language) => {
+    createPage({
+      path: language === 'en' ? '/' : `/${language}/index.html`,
+      component: frontPageTemplate,
+      context: { id: '', lng: language, sideBar, title: '', idTitleMap, markDownId: '/' },
     });
   });
 };
