+++ website/package.json
@@ -15,7 +15,7 @@
     "@mikaelkristiansson/domready": "1.0.10",
     "@reach/router": "1.3.4",
     "babel-preset-gatsby": "^0.4.12",
-    "clsx": "^1.1.1",
+    "classnames": "^2.2.6",
     "emotion-theming": "10.0.27",
     "event-source-polyfill": "^1.0.17",
     "fontsource-roboto": "^2.2.6",
